# Optimized std.datetime.systime.SysTime.fromISOExtString