```
old		12 secs, 875 ms, 52 μs, and 1 hnsec
new		6 secs, 374 ms, 348 μs, and 1 hnsec
```

code

<details>

```d
import std.stdio;
import std.exception;
import std.conv;
import std.ascii;
import std.range;
import std.traits;
import std.typecons;
import std.string;
import std.datetime.systime;
import std.datetime.date;
import std.datetime.timezone;
import std.datetime.stopwatch;
import core.time;
import std.random;

enum testCount = 5_000_000;

__gshared a = "2013-02-07T04:39:37.000050392";

SysTime fromISOExtString1(S)(in S isoExtString, immutable TimeZone tz = null) @safe
        if (isSomeString!(S))
{
    import std.algorithm.searching : countUntil, find;
    import std.conv : to;
    import std.string : strip;

    auto dstr = to!dstring(strip(isoExtString));

    auto tIndex = dstr.countUntil('T');
    enforce(tIndex != -1, new DateTimeException(format("Invalid ISO Extended String: %s", isoExtString)));

    auto found = dstr[tIndex + 1 .. $].find('.', 'Z', '+', '-');
    auto dateTimeStr = dstr[0 .. $ - found[0].length];

    dstring fracSecStr;
    dstring zoneStr;

    if (found[1] != 0)
    {
        if (found[1] == 1)
        {
            auto foundTZ = found[0].find('Z', '+', '-');

            if (foundTZ[1] != 0)
            {
                fracSecStr = found[0][0 .. $ - foundTZ[0].length];
                zoneStr = foundTZ[0];
            }
            else
                fracSecStr = found[0];
        }
        else
            zoneStr = found[0];
    }

    try
    {
        auto dateTime = DateTime.fromISOExtString(dateTimeStr);
        auto fracSec = fracSecsFromISOString(fracSecStr);
        Rebindable!(immutable TimeZone) parsedZone;

        if (zoneStr.empty)
            parsedZone = LocalTime();
        else if (zoneStr == "Z")
            parsedZone = UTC();
        else
            parsedZone = fromISOExtString3(zoneStr);

        auto retval = SysTime(dateTime, fracSec, parsedZone);

        if (tz !is null)
            retval.timezone = tz;

        return retval;
    }
    catch (DateTimeException dte)
        throw new DateTimeException(format("Invalid ISO Extended String: %s", isoExtString));
}

SysTime fromISOExtString2(S)(in S isoExtString, immutable TimeZone tz = null) @safe
        if (isSomeString!(S))
{
    import std.algorithm.searching : countUntil, find;
    import std.conv : to;
    import std.string : strip, indexOf;

    auto str = strip(isoExtString);

    auto tIndex = str.indexOf('T');
    enforce(tIndex != -1, new DateTimeException(format("Invalid ISO Extended String: %s", isoExtString)));

    auto found = str[tIndex + 1 .. $].byCodeUnit.find('.', 'Z', '+', '-');
    auto dateTimeStr = str[0 .. $ - found[0].length];

    typeof(str) fracSecStr;
    typeof(str) zoneStr;

    if (found[1] != 0)
    {
        if (found[1] == 1)
        {
            auto foundTZ = found[0].find('Z', '+', '-');

            if (foundTZ[1] != 0)
            {
                static if (isNarrowString!S)
                {
                    fracSecStr = found[0][0 .. $ - foundTZ[0].length].source;
                    zoneStr = foundTZ[0].source;
                }
                else
                {
                    fracSecStr = found[0][0 .. $ - foundTZ[0].length];
                    zoneStr = foundTZ[0];
                }
            }
            else
            {
                static if (isNarrowString!S)
                    fracSecStr = found[0].source;
                else
                    fracSecStr = found[0];
            }
        }
        else
        {
            static if (isNarrowString!S)
                zoneStr = found[0].source;
            else
                zoneStr = found[0];
        }
    }

    try
    {
        auto dateTime = DateTime.fromISOExtString(dateTimeStr);
        auto fracSec = fracSecsFromISOString(fracSecStr);
        Rebindable!(immutable TimeZone) parsedZone;

        if (zoneStr.empty)
            parsedZone = LocalTime();
        else if (zoneStr == "Z")
            parsedZone = UTC();
        else
            parsedZone = fromISOExtString3(zoneStr);

        auto retval = SysTime(dateTime, fracSec, parsedZone);

        if (tz !is null)
            retval.timezone = tz;

        return retval;
    }
    catch (DateTimeException dte)
        throw new DateTimeException(format("Invalid ISO Extended String: %s", isoExtString));
}

Duration fracSecsFromISOString(S)(in S isoString) @trusted pure
if (isSomeString!S)
{
    import std.algorithm.searching : all;
    import std.ascii : isDigit;
    import std.conv : to;
    import std.string : representation;
    import core.time;

    if (isoString.empty)
        return Duration.zero;

    auto str = isoString.representation;

    enforce(str[0] == '.', new DateTimeException("Invalid ISO String"));
    str.popFront();

    enforce(!str.empty && all!isDigit(str), new DateTimeException("Invalid ISO String"));

    dchar[7] fullISOString = void;
    foreach (i, ref dchar c; fullISOString)
    {
        if (i < str.length)
            c = str[i];
        else
            c = '0';
    }

    return hnsecs(to!int(fullISOString[]));
}

immutable(SimpleTimeZone) fromISOExtString3(S)(S isoExtString) @safe pure
        if (isSomeString!S)
{
    import std.algorithm.searching : startsWith;
    import std.conv : ConvException, to;
    import std.datetime.date : DateTimeException;
    import std.exception : enforce;
    import std.format : format;
    import std.string : indexOf;

    auto whichSign = isoExtString.startsWith('-', '+');
    enforce!DateTimeException(whichSign > 0, format("Invalid ISO String: %s", isoExtString));
    auto sign = whichSign == 1 ? -1 : 1;

    isoExtString = isoExtString[1 .. $];
    enforce!DateTimeException(!isoExtString.empty, format("Invalid ISO String: %s", isoExtString));

    immutable colon = isoExtString.indexOf(':');
    S hoursStr;
    S minutesStr;
    int hours, minutes;

    if (colon != -1)
    {
        hoursStr = isoExtString[0 .. colon];
        minutesStr = isoExtString[colon + 1 .. $];
        enforce!DateTimeException(minutesStr.length == 2, format("Invalid ISO String: %s", isoExtString));
    }
    else
    {
        hoursStr = isoExtString;
    }

    enforce!DateTimeException(hoursStr.length == 2, format("Invalid ISO String: %s", isoExtString));

    try
    {
        // cast to int from uint is used because it checks for
        // non digits without extra loops
        hours = cast(int) to!uint(hoursStr);
        minutes = cast(int) (minutesStr.empty ? 0 : to!uint(minutesStr));
    }
    catch (ConvException)
    {
        throw new DateTimeException(format("Invalid ISO String: %s", isoExtString));
    }

    enforce!DateTimeException(hours < 24 && minutes < 60, format("Invalid ISO String: %s", isoExtString));

    return new immutable SimpleTimeZone(sign * (dur!"hours"(hours) + dur!"minutes"(minutes)));
}

void main()
{
    SysTime res;

    auto result = to!Duration(benchmark!(() => res = fromISOExtString1(a))(testCount)[0]); // 16000000
    auto result2 = to!Duration(benchmark!(() => res = fromISOExtString2(a))(testCount)[0]); // 32000000

    writeln(res); // force execution

    writeln("old", "\t\t", result);
    writeln("new", "\t\t", result2);
}
```

</details>